Move exercise related files to exercise package in app module
