+++ app/src/main/java/com/lukasz/witkowski/training/planner/navigation/Navigation.kt
@@ -13,15 +13,15 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import androidx.navigation.navigation
-import com.lukasz.witkowski.training.planner.exercise.CreateExerciseViewModel
-import com.lukasz.witkowski.training.planner.exercise.ExercisesListViewModel
+import com.lukasz.witkowski.training.planner.exercise.createExercise.CreateExerciseViewModel
+import com.lukasz.witkowski.training.planner.exercise.exercisesList.ExercisesListViewModel
 import com.lukasz.witkowski.training.planner.training.CreateTrainingViewModel
 import com.lukasz.witkowski.training.planner.training.TrainingsListViewModel
 import com.lukasz.witkowski.training.planner.ui.*
-import com.lukasz.witkowski.training.planner.ui.createExercise.CreateExerciseScreen
+import com.lukasz.witkowski.training.planner.exercise.createExercise.CreateExerciseScreen
 import com.lukasz.witkowski.training.planner.ui.createTraining.CreateTrainingScreen
 import com.lukasz.witkowski.training.planner.ui.createTraining.PickExerciseScreen
-import com.lukasz.witkowski.training.planner.ui.exercisesList.ExercisesScreen
+import com.lukasz.witkowski.training.planner.exercise.exercisesList.ExercisesScreen
 import com.lukasz.witkowski.training.planner.ui.trainingOverview.TrainingOverviewScreen
 import com.lukasz.witkowski.training.planner.ui.trainingOverview.TrainingOverviewViewModel
 import com.lukasz.witkowski.training.planner.ui.trainingsList.TrainingsScreen
