+++ app/src/main/java/com/lukasz/witkowski/training/planner/ui/BaseCategoryFilteredListViewModel.kt
@@ -1,7 +1,7 @@
 package com.lukasz.witkowski.training.planner.ui
 
 import androidx.lifecycle.ViewModel
-import com.lukasz.witkowski.training.planner.exercise.Category
+import com.lukasz.witkowski.training.planner.exercise.models.Category
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
