+++ app/src/main/java/com/lukasz/witkowski/training/planner/ui/trainingsList/TrainingsScreen.kt
@@ -25,16 +25,15 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.lukasz.witkowski.training.planner.R
-import com.lukasz.witkowski.training.planner.exercise.Category
-import com.lukasz.witkowski.training.planner.exercise.CategoryMapper
-import com.lukasz.witkowski.training.planner.exercise.allCategories
+import com.lukasz.witkowski.training.planner.exercise.models.CategoryMapper
 import com.lukasz.witkowski.training.planner.exercise.domain.ExerciseCategory
 import com.lukasz.witkowski.training.planner.training.domain.TrainingPlan
 import com.lukasz.witkowski.training.planner.training.TrainingsListViewModel
 import com.lukasz.witkowski.training.planner.ui.components.CategoryChip
 import com.lukasz.witkowski.training.planner.ui.components.ListCardItem
 import com.lukasz.witkowski.training.planner.ui.components.NoDataMessage
-import com.lukasz.witkowski.training.planner.ui.exercisesList.CategoryFilters
+import com.lukasz.witkowski.training.planner.exercise.exercisesList.CategoryFilters
+import com.lukasz.witkowski.training.planner.exercise.models.allCategories
 
 @Composable
 fun TrainingsScreen(
